// app/src/models/models.ts
export type Track = {
  addedAt: Date;
  albumTitle: string;
  albumArtists: string;
  artists: string;
  bitrate: number;
  cloudStatus?: "Apple Music" | "Matched" | "Uploaded";
  contentRating?: "Clean" | "Explicit";
  comments?: string;
  composers: string;
  disc: number;
  duration: number;
  fileFormat: string;
  fileSize: number;
  gapless: boolean;
  genre: string;
  likeFactor: 1 | -1 | 0;
  modifiedAt: Date;
  normalization: number;
  persistentId: string;
  playCount: number;
  rating?: number;
  releaseDate?: Date;
  releaseYear: number;
  sampleRate?: number;
  skipCount: number;
  title: string;
  track: number;
  volumeAdjustment?: number;
  work?: string;
};